+++ apps/mobile/src/hooks/use-multi-tap-reveal.ts
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import { useMultiTapAction } from "~/hooks/use-multi-tap-action";
 
 interface MultiTapRevealOptions {
   tapsRequired?: number;
@@ -11,25 +12,11 @@ interface MultiTapRevealResult {
 }
 
 export function useMultiTapReveal(options?: MultiTapRevealOptions): MultiTapRevealResult {
-  const tapsRequired = options?.tapsRequired ?? 4;
-  const intervalMs = options?.intervalMs ?? 600;
-
   const [isVisible, setIsVisible] = useState(false);
-  const [tapCount, setTapCount] = useState(0);
-  const [lastTapTime, setLastTapTime] = useState(0);
-
-  function handleTap() {
-    const now = Date.now();
-    const withinWindow = now - lastTapTime < intervalMs;
-    const nextCount = withinWindow ? tapCount + 1 : 1;
-    setTapCount(nextCount);
-    setLastTapTime(now);
-    if (nextCount >= tapsRequired) {
-      setIsVisible(true);
-      setTapCount(0);
-      setLastTapTime(0);
-    }
-  }
+  const handleTap = useMultiTapAction(() => setIsVisible(true), {
+    tapsRequired: options?.tapsRequired ?? 4,
+    intervalMs: options?.intervalMs,
+  });
 
   return { isVisible, handleTap };
 }
